+++ src/rendering/renderer/fence.h
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include "dxr_includes.h"
+#include "rendering/dxr_includes.h"
 
 class Fence
 {
